+++ src/Endpoints/Indices.php
@@ -5,8 +5,8 @@ namespace MarketDataApp\Endpoints;
 use Carbon\Carbon;
 use GuzzleHttp\Exception\GuzzleException;
 use MarketDataApp\Client;
-use MarketDataApp\Endpoints\Responses\IndicesCandles;
-use MarketDataApp\Endpoints\Responses\IndicesQuote;
+use MarketDataApp\Endpoints\Responses\Indices\Candles;
+use MarketDataApp\Endpoints\Responses\Indices\Quote;
 
 class Indices
 {
@@ -26,9 +26,9 @@ class Indices
      * do not use $DJI, ^DJI, .DJI, DJI.X, etc.
      * @throws GuzzleException
      */
-    public function quote(string $symbol): IndicesQuote
+    public function quote(string $symbol): Quote
     {
-        return new IndicesQuote($this->client->execute(self::BASE_URL . "quotes/{$symbol}"));
+        return new Quote($this->client->execute(self::BASE_URL . "quotes/{$symbol}"));
     }
 
     /**
@@ -37,9 +37,7 @@ class Indices
      * @param string $symbol The index symbol, without any leading or trailing index identifiers. For example, use DJI
      * do not use $DJI, ^DJI, .DJI, DJI.X, etc.
      * @param Carbon $from The leftmost candle on a chart (inclusive). If you use countback, to is not required.
-     * Accepted timestamp inputs: ISO 8601, unix, spreadsheet.
-     * @param Carbon|null $to The rightmost candle on a chart (inclusive). Accepted timestamp inputs: ISO 8601, unix,
-     * spreadsheet.
+     * @param Carbon|null $to The rightmost candle on a chart (inclusive).
      * @param string $resolution The duration of each candle.
      * Minutely Resolutions: (minutely, 1, 3, 5, 15, 30, 45, ...) Hourly Resolutions: (hourly, H, 1H, 2H, ...)
      * Daily Resolutions: (daily, D, 1D, 2D, ...)
@@ -48,7 +46,7 @@ class Indices
      * Yearly Resolutions:(yearly, Y, 1Y, 2Y, ...)
      * @param int|null $countback Will fetch a number of candles before (to the left of) to. If you use from, countback
      * is not required.
-     * @return IndicesCandles
+     * @return Candles
      * @throws GuzzleException
      */
     public function candles(
@@ -57,8 +55,8 @@ class Indices
         Carbon $to = null,
         string $resolution = 'D',
         int $countback = null
-    ): IndicesCandles {
-        return new IndicesCandles($this->client->execute(self::BASE_URL . "candles/{$resolution}/{$symbol}/",
+    ): Candles {
+        return new Candles($this->client->execute(self::BASE_URL . "candles/{$resolution}/{$symbol}/",
             compact('from', 'to', 'countback')
         ));
     }