+++ src/Endpoints/Stocks.php
@@ -5,9 +5,10 @@ namespace MarketDataApp\Endpoints;
 use Carbon\Carbon;
 use GuzzleHttp\Exception\GuzzleException;
 use MarketDataApp\Client;
-use MarketDataApp\Endpoints\Responses\StocksBulkQuotes;
-use MarketDataApp\Endpoints\Responses\StocksQuote;
-use MarketDataApp\Endpoints\Responses\StocksQuotes;
+use MarketDataApp\Endpoints\Responses\Stocks\BulkQuotes;
+use MarketDataApp\Endpoints\Responses\Stocks\Candles;
+use MarketDataApp\Endpoints\Responses\Stocks\Quote;
+use MarketDataApp\Endpoints\Responses\Stocks\Quotes;
 
 class Stocks
 {
@@ -20,6 +21,35 @@ class Stocks
         $this->client = $client;
     }
 
+        /**
+     * Get historical price candles for an index.
+     *
+     * @param string $symbol The company's ticker symbol.
+     * @param Carbon $from The leftmost candle on a chart (inclusive). If you use countback, to is not required.
+     * @param Carbon|null $to The rightmost candle on a chart (inclusive).
+     * @param string $resolution The duration of each candle.
+     * Minutely Resolutions: (minutely, 1, 3, 5, 15, 30, 45, ...) Hourly Resolutions: (hourly, H, 1H, 2H, ...)
+     * Daily Resolutions: (daily, D, 1D, 2D, ...)
+     * Weekly Resolutions: (weekly, W, 1W, 2W, ...)
+     * Monthly Resolutions: (monthly, M, 1M, 2M, ...)
+     * Yearly Resolutions:(yearly, Y, 1Y, 2Y, ...)
+     * @param int|null $countback Will fetch a number of candles before (to the left of) to. If you use from, countback
+     * is not required.
+     * @return Candles
+     * @throws GuzzleException
+     */
+    public function candles(
+        string $symbol,
+        Carbon $from,
+        Carbon $to = null,
+        string $resolution = 'D',
+        int $countback = null
+    ): Candles {
+        return new Candles($this->client->execute(self::BASE_URL . "candles/{$resolution}/{$symbol}/",
+            compact('from', 'to', 'countback')
+        ));
+    }
+
     /**
      * Get a real-time price quote for a stock.
      *
@@ -28,9 +58,9 @@ class Stocks
      * default this parameter is false if omitted.
      * @throws GuzzleException
      */
-    public function quote(string $symbol, bool $fifty_two_week = false): StocksQuote
+    public function quote(string $symbol, bool $fifty_two_week = false): Quote
     {
-        return new StocksQuote($this->client->execute(self::BASE_URL . "quotes/{$symbol}",
+        return new Quote($this->client->execute(self::BASE_URL . "quotes/{$symbol}",
             ['52week' => $fifty_two_week]));
     }
 
@@ -42,7 +72,7 @@ class Stocks
      * default this parameter is false if omitted.
      * @throws \Throwable
      */
-    public function quotes(array $symbols, bool $fifty_two_week = false): StocksQuotes|StocksBulkQuotes
+    public function quotes(array $symbols, bool $fifty_two_week = false): Quotes|BulkQuotes
     {
         // Execute standard quotes in parallel
         $calls = [];
@@ -50,7 +80,7 @@ class Stocks
             $calls[] = ["/stocks/quotes/$symbol", ['52week' => $fifty_two_week]];
         }
 
-        return new StocksQuotes($this->client->executeInParallel($calls));
+        return new Quotes($this->client->executeInParallel($calls));
     }
 
     /**
@@ -67,13 +97,13 @@ class Stocks
      * @throws GuzzleException
      * @throws \Exception
      */
-    public function bulkQuotes(array $symbols = [], bool $snapshot = false): StocksBulkQuotes
+    public function bulkQuotes(array $symbols = [], bool $snapshot = false): BulkQuotes
     {
         if (empty($symbols) && !$snapshot) {
             throw new \InvalidArgumentException('Either symbols or snapshot must be set');
         }
 
-        return new StocksBulkQuotes($this->client->execute(self::BASE_URL . "bulkquotes",
+        return new BulkQuotes($this->client->execute(self::BASE_URL . "bulkquotes",
             ['symbols' => implode(',', $symbols), 'snapshot' => $snapshot]));
     }
 }