+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 # Changelog
 
-All notable changes to `marketdataapp-sdk-php` will be documented in this file.
+## v0.2.0-alpha
 
+- Completed Indices endpoints.
+- Completed Stocks endpoints.
+- Added stubs for remaining endpoints.
+
+## v0.1.0-alpha
+
+- Initial release with single Indices > quote endpoint.